requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
autopep8==1.5.4
cffi==1.14.2
dj-database-url==0.5.0
Django==2.2.15
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-js-asset==1.2.2
django-modelcluster==5.0.2
django-mptt==0.11.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
